requirements.txt
streamlit
scikit-learn
pandas
tmdbv3api
numpy
requests
python-dotenv
